requirements-lowend-cpu.txt
click
mypy[mypyc]
numpy
pip
polars-lts-cpu
py-cpuinfo
setuptools
wheel
zstandard
